# docs/api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EduMark",
    "description": "\n基于大模型的中小学作业批改系统，提供以下核心能力：\n\n- **作业提交与批改**：上传作业图片，自动识别题目并进行 AI 批改\n- **知识库管理**：上传教学资料，构建知识库辅助批改\n- **用户管理**：学生/老师/管理员三种角色，权限分级\n- **模型管理**：配置视觉模型和语言模型\n- **配置管理**：个性化识别模式与增强选项\n",
    "contact": {
      "name": "EduMark Team",
      "email": "admin@edumark.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/homework/submit": {
      "post": {
        "tags": [
          "作业管理"
        ],
        "summary": "提交作业",
        "description": "上传作业图片进行批改，异步处理。\n\n- 支持一次上传多张图片\n- 返回 task_id 用于后续查询批改进度和结果\n- 仅学生角色可提交",
        "operationId": "submit_homework_api_v1_homework_submit_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_submit_homework_api_v1_homework_submit_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "提交成功，返回 task_id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_SubmitResponse_"
                }
              }
            }
          },
          "403": {
            "description": "仅学生角色可提交"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/homework/task/page": {
      "post": {
        "tags": [
          "作业管理"
        ],
        "summary": "分页查询任务",
        "description": "分页查询作业任务列表。\n\n- 学生只能查看自己的任务\n- 管理员可查看所有任务\n- 支持排序和过滤条件",
        "operationId": "page_task_api_v1_homework_task_page_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PageDTO"
              },
              "examples": {
                "默认分页": {
                  "summary": "默认分页查询（第1页，每页10条）",
                  "value": {
                    "page_num": 1,
                    "page_size": 10
                  }
                },
                "按状态过滤": {
                  "summary": "筛选已完成的任务",
                  "value": {
                    "page_num": 1,
                    "page_size": 10,
                    "query": {
                      "status": "done"
                    }
                  }
                },
                "按创建时间排序": {
                  "summary": "按创建时间降序排列",
                  "value": {
                    "page_num": 1,
                    "page_size": 10,
                    "sort_fields": [
                      {
                        "field": "create_time",
                        "direction": "desc"
                      }
                    ]
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "查询成功，返回分页数据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_PageVO_TaskOut__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/homework/status/{task_id}": {
      "get": {
        "tags": [
          "作业管理"
        ],
        "summary": "查询批改进度",
        "description": "查询作业批改的当前状态。\n\n- 只能查看自己的任务或管理员可查看所有\n- 状态值：`pending`（待处理）、`processing`（处理中）、`done`（已完成）",
        "operationId": "get_status_api_v1_homework_status__task_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "任务唯一标识",
              "title": "Task Id"
            },
            "description": "任务唯一标识"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功，返回任务状态",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_StatusResponse_"
                }
              }
            }
          },
          "403": {
            "description": "无权查看该任务"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/homework/result/{task_id}": {
      "get": {
        "tags": [
          "作业管理"
        ],
        "summary": "查询批改结果",
        "description": "获取完整的批改结果，包含：\n\n- 原始作业图片（带访问链接）\n- 每道题目的识别文本、学生答案\n- 每道题目的裁切区域坐标\n- AI 批改得分、评语、错因分析\n- 关联的知识库引用",
        "operationId": "get_result_api_v1_homework_result__task_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "任务唯一标识",
              "title": "Task Id"
            },
            "description": "任务唯一标识"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功，返回完整批改结果",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_HomeworkResult_"
                }
              }
            }
          },
          "403": {
            "description": "无权查看该任务"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/homework/task/{task_id}": {
      "delete": {
        "tags": [
          "作业管理"
        ],
        "summary": "删除作业",
        "description": "删除作业及相关题目、批改结果。\n\n- 只能删除自己的作业\n- 管理员可删除任意作业\n- 级联删除所有关联数据",
        "operationId": "delete_task_api_v1_homework_task__task_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "任务唯一标识",
              "title": "Task Id"
            },
            "description": "任务唯一标识"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          },
          "403": {
            "description": "无权删除该作业"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_submit_homework_api_v1_homework_submit_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files",
            "description": "作业图片文件，支持多张"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject",
            "description": "科目，如语文、数学、英语"
          },
          "grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade",
            "description": "年级，如一年级、二年级"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_submit_homework_api_v1_homework_submit_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HomeworkResult": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id",
            "description": "任务唯一标识"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "任务状态"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject",
            "description": "科目"
          },
          "grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade",
            "description": "年级"
          },
          "mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode",
            "description": "批改模式"
          },
          "create_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Create Time",
            "description": "创建时间"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/ImageInfo"
            },
            "type": "array",
            "title": "Images",
            "description": "作业原始图片列表"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionResult"
            },
            "type": "array",
            "title": "Questions",
            "description": "识别出的题目及批改结果列表"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "status"
        ],
        "title": "HomeworkResult",
        "example": {
          "create_time": "2026-06-04T10:00:00",
          "grade": "二年级",
          "images": [
            {
              "url": "https://minio.example.com/homework/1/1717488000_math_homework.jpg"
            }
          ],
          "mode": "aliyun",
          "questions": [
            {
              "blocks": [
                {
                  "url": "https://minio.example.com/homework/1/blocks/1717488000_q1.jpg",
                  "x1": 100.0,
                  "x2": 500.0,
                  "y1": 50.0,
                  "y2": 200.0
                }
              ],
              "correction": {
                "analysis": "12-3+5=14，先减后加，运算顺序正确",
                "comment": "回答正确，计算过程无误",
                "result": "correct",
                "score": 10.0
              },
              "create_time": "2026-06-04T10:01:00",
              "knowledge_refs": [
                {
                  "content": "加减混合运算：在没有括号的算式里，只有加减法，要从左往右按顺序计算。",
                  "knowledge_id": 1,
                  "score": 0.95,
                  "title": "二年级数学知识点汇总.pdf"
                }
              ],
              "no": "1",
              "question_text": "小明有12个苹果，吃了3个，又买了5个，现在有多少个苹果？",
              "question_type": "应用题",
              "student_answer": "14个"
            }
          ],
          "status": "completed",
          "subject": "数学",
          "task_id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        }
      },
      "ImageInfo": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url",
            "description": "作业图片访问地址"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "ImageInfo",
        "example": {
          "url": "https://minio.example.com/homework/1/1717488000_math_homework.jpg"
        }
      },
...
      "PageDTO": {
        "properties": {
          "page_num": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page Num",
            "description": "页码，从 1 开始",
            "default": 1
          },
          "page_size": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "description": "每页条数，范围 1~100",
            "default": 10
          },
          "sort_fields": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SortField"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Fields",
            "description": "排序规则列表，可选"
          },
          "query": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query",
            "description": "查询过滤条件，键值对形式，可选"
          }
        },
        "type": "object",
        "title": "PageDTO"
      },
...
      "PageVO_TaskOut_": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "总记录数"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/TaskOut"
            },
            "type": "array",
            "title": "Rows",
            "description": "当前页数据列表"
          }
        },
        "type": "object",
        "required": [
          "total",
          "rows"
        ],
        "title": "PageVO[TaskOut]",
        "example": {
          "rows": [],
          "total": 100
        }
      },
...
      "Result": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码，1=成功，-1=失败"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "提示信息，成功时为 'success'，失败时为错误描述"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "响应数据，成功时返回业务数据，失败时为 null"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Result",
        "example": {
          "code": 1,
          "message": "success"
        }
      },
...
      "Result_HomeworkResult_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码，1=成功，-1=失败"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "提示信息，成功时为 'success'，失败时为错误描述"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HomeworkResult"
              },
              {
                "type": "null"
              }
            ],
            "description": "响应数据，成功时返回业务数据，失败时为 null"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Result[HomeworkResult]",
        "example": {
          "code": 1,
          "message": "success"
        }
      },
...
      "Result_PageVO_TaskOut__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码，1=成功，-1=失败"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "提示信息，成功时为 'success'，失败时为错误描述"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PageVO_TaskOut_"
              },
              {
                "type": "null"
              }
            ],
            "description": "响应数据，成功时返回业务数据，失败时为 null"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Result[PageVO[TaskOut]]",
        "example": {
          "code": 1,
          "message": "success"
        }
      },
...
      "Result_StatusResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码，1=成功，-1=失败"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "提示信息，成功时为 'success'，失败时为错误描述"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StatusResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "响应数据，成功时返回业务数据，失败时为 null"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Result[StatusResponse]",
        "example": {
          "code": 1,
          "message": "success"
        }
      },
      "Result_SubmitResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码，1=成功，-1=失败"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "提示信息，成功时为 'success'，失败时为错误描述"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SubmitResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "响应数据，成功时返回业务数据，失败时为 null"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Result[SubmitResponse]",
        "example": {
          "code": 1,
          "message": "success"
        }
      },
...
      "SortDirection": {
        "type": "string",
        "enum": [
          "asc",
          "desc"
        ],
        "title": "SortDirection"
      },
      "SortField": {
        "properties": {
          "field": {
            "type": "string",
            "minLength": 1,
            "title": "Field",
            "description": "排序字段名"
          },
          "direction": {
            "$ref": "#/components/schemas/SortDirection",
            "description": "排序方向，asc=升序，desc=降序",
            "default": "asc"
          }
        },
        "type": "object",
        "required": [
          "field"
        ],
        "title": "SortField"
      },
      "StatusResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id",
            "description": "任务唯一标识"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "当前状态：pending=待处理，processing=处理中，completed=已完成，failed=失败"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "status"
        ],
        "title": "StatusResponse",
        "example": {
          "status": "completed",
          "task_id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        }
      },
      "SubmitResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id",
            "description": "提交成功后分配的任务唯一标识"
          }
        },
        "type": "object",
        "required": [
          "task_id"
        ],
        "title": "SubmitResponse",
        "example": {
          "task_id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        }
      },
      "TaskOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "任务记录 ID"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id",
            "description": "任务唯一标识"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "所属用户 ID"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject",
            "description": "科目"
          },
          "grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade",
            "description": "年级"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "description": "批改模式"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "任务状态"
          },
          "create_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Create Time",
            "description": "创建时间"
          },
          "update_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Update Time",
            "description": "最后更新时间"
          }
        },
        "type": "object",
        "required": [
          "id",
          "task_id",
          "user_id",
          "mode",
          "status"
        ],
        "title": "TaskOut",
        "example": {
          "create_time": "2026-06-04T10:00:00",
          "grade": "二年级",
          "id": 1,
          "mode": "aliyun",
          "status": "completed",
          "subject": "数学",
          "task_id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
          "update_time": "2026-06-04T10:05:00",
          "user_id": 1
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  },
...
}